some minor bug fixes
Cancel out some CLI chars that will break the stream feed commands,
Chars are: White space in video path, -, ', (, ), &

Also fixed an issue where the ps string had a space before it breaking
the feed process killer break.

Moved the entire URL for the video player to the PHP script so that it
can be compiled on the fly.

Added process checker for the index page to check if the ffserver and
tongue servers are running or not.