new upload
The seedlings of the Documentation are in Docs/
Renamed the daemon and webUI folders to TongueD and Tongue_WebUI
Movies part of the project is forming along. Daemon searches the movies
folder and can distinguish between single movie files, 2 part CD movie
splits, and Raw DVD VIDEO_TS folders.
Two part and DVD VIDEO_TS folder streaming is not working, but the
single video file should work, although I have not tested it yet.
WebUI for the movies section is started, movies are listed out, when you
click on one a details page comes up on the right and shows you the
details of the movie ( raw Dvd, multi part movie, runtime, dimentions)