more work on the new feed mgmt system
Made a multi-dimensional dictionary search function, its really a hack
job and a half but it works. it returns a list of values that are the
path to the value that you are searching for.
Used this to search the available streams dict to find the stream/feed
that is being used.
Still need to make it actually manage them but at least the data is
being organized.